// MyAuctionClientApp/src/types/Vehicle.ts
export interface Vehicle {
    vehicleId: number
    brandAndModel: string
    color: string
    manufacturingYear: number
    engineCapacity: number
    price: number
    milage: number
    plateNumber: string
    auctionPrice: number
    additionalInformation: string
    startTime: string
    endTime: string
    isActive: boolean
    image: string
    sellerId: string
    bids: any
  }
  